Updating random generators to use numpy, to maintain bit generator (#97)
After a change to re-initialize the fast-binomial generators when their parameters change, the population dynamics started to break when we switched timesteps.

This is because when you re-initialize a random generator with the same seed, the baseline values for the bit-generator will be the same. 

- In the fast-binomial package the generators are initialized typically with a seed, and a probability. However, even if the probability is different, as long as the seed is the same, the "bit-generated values" will be produced in the same sequence as the first initialization, which results in the randomness of the "newly initialized" generator being correlated with the old generator. 
  - This causes people with a higher chance of death before, to have a higher chance again, regardless of their current age. 
- Another issue with fast_binomial is that if you pass in the probability in a specific random call, rather than at the beginning, it recreates the bit generator, which again, means the values will be correlated.

The fix chosen was to switch to using numpy, to maintain the same bit-generator across multiple calls with the same seed. It is a cleaner solution, and an already existing one. Subjectively, no performance hits were noticed.