Dynamically sizing delay arrays when timesteps change (#98)
When timesteps are changed during the model run, the model called "Delay Arrays" break. In summary, the delay arrays work by creating a rows for each individual for each timestep that we need to delay certain lifecycle elements in the model. This includes new worms entering humans (takes 10 months), mf developing in blackflies (4 days) and exposure to blackfly bites (4 days). With a timestep of 1 day, this would give us matrices of 280(days) x 400 (people), 4x400, and 4x400. Currently, the number of rows is determined by the initial timestep, and then never touched again. That means if we start at a longer timestep (i.e 7 days), the matrices will be initialized to 40x400, 1x400, and 1x400. When we switch back to a 1 day timestep, these dimensions stay, even though they need to be adjusted to the size determined by the new timestep. The fix is to up- and down-scales these arrays appropriately to make sure that the number of rows in the matrices will match the timestep.